opt: remove redundant Struct
