+++ include/types.h
@@ -40,3 +40,17 @@ enum PowerMeterResponse {
 };
 
 #endif // STATUS_H
+
+#ifndef __RTCDateTime__                                                                                       // nếu mảng lưu thời gian chưa được tạo
+#define __RTCDateTime__                                                                                       // đánh dấu đã tạo
+struct RTCDateTime {                                                                                          // mảng dữ liệu lưu thời gian
+  uint16_t year     = 0;                                                                                      // năm
+  uint8_t month     = 0;                                                                                      // tháng
+  uint8_t day       = 0;                                                                                      // ngày
+  uint8_t hour      = 0;                                                                                      // giờ
+  uint8_t minute    = 0;                                                                                      // phút
+  uint8_t second    = 0;                                                                                      // giây
+  uint8_t dayOfWeek = 0;                                                                                      // thứ
+  uint32_t unixtime = 0;                                                                                      // thời gian dài tính bằng giây từ ngày 1/1/1970
+};                                                                                                            //
+#endif
